Put UI on separate thread
This entails a significant change to the program architecture.
Instead of the main method owning the UI object and calling its
methods, it now simply spawns the new thread, and communicates with
the UI via message passing (in both directions). This allows
long-running tasks to happen in the background without locking up
the UI.