Fix two bugs with marking episodes played/unplayed
Stupid typo led to panic for marking episodes in the 0th index.
Also will now refresh title of podcast to update unplayed totals
when one is marked.

In addition to this, this commit pulls the UI module apart a bit
because the ui.rs file was getting to be a monstrosity.