Add updated version of crate and change TLS support
The `rfc822_sanitizer` crate appears to be unmaintained, but uses
older, unstable versions of a couple crates; I have cloned the
crate to update the versions, which cuts out a few additional
dependencies from shellcaster.

The `rustls` crate seems to be failing on a couple websites for me,
so I've changed the default to `native-tls`, but added a feature
to be able to switch between them if so desired.