Major refactor to use item IDs instead of indexes
Includes a change in the LockVec structure to include a HashMap
to identify podcasts/episodes by ID, plus an additional vector
that tracks the order of the items. This is somewhat more bulky,
but now allows the app to pass around IDs (which are invariant)
rather than indexes (which may change if new podcasts or episodes
are added, removed, etc.).