Add more error handling
Adds anyhow as project dependency, and changes many functions to
return anyhow::Result, which reduces many of the unwrap() calls.
There are still many places where the code could potentially panic,
but the majority of the remaining unwrap() calls are at least
changed to expect() to provide a message as well.